# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nutrition API",
    "version": "1.0.0",
    "description": "API for tracking daily calorie intake and managing user data"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/users/current": {
      "get": {
        "summary": "Get current logged-in user",
        "description": "Fetches the currently authenticated user from the system.",
        "operationId": "getCurrentUser",
        "tags": ["Auth"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched the current user details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The ID of the user"
                    },
                    "username": {
                      "type": "string",
                      "description": "The username of the user"
                    },
                    "email": {
                      "type": "string",
                      "description": "The email of the user"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "The date and time when the user was created"
                    },
                    "example": {
                      "id": "123",
                      "username": "user123",
                      "email": "user@example.com",
                      "createdAt": "2024-01-01T12:00:00Z"
                    }
                  },
                  "required": ["id", "username", "email", "createdAt"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. The user is not authenticated."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/api/public/{height}/{desiredWeight}/{age}/{bloodGroupIndex}/{currentWeight}": {
      "get": {
        "summary": "Get general recommended daily calories intake based on height, desiredWeight, age, currentWeight and returnes forbiden  categories of aliments to eat based on blood group",
        "description": "This endpoint calculates the recommended daily calorie intake based on height, desired weight, current weight, and age, and retrieves health categories based on the blood group index. It checks if the given blood group is allowed and fetches categories accordingly.",
        "operationId": "getCategoriesForBloodGroup",
        "parameters": [
          {
            "name": "height",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 170
            },
            "description": "The height of the individual in cm."
          },
          {
            "name": "desiredWeight",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 70
            },
            "description": "The desired weight of the individual in kg."
          },
          {
            "name": "age",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 30
            },
            "description": "The age of the individual."
          },
          {
            "name": "bloodGroupIndex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2
            },
            "description": "The blood group index for fetching the relevant categories."
          },
          {
            "name": "currentWeight",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 75
            },
            "description": "The current weight of the individual in kg."
          }
        ],
        "tags": ["Nutrition"],
        "responses": {
          "200": {
            "description": "Successfully retrieved daily calorie intake and categories for the blood group.",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "recommendedDailyCaloriesIntake": 2200,
                  "data": [
                    { "categoryId": 1, "categoryName": "Category 1" },
                    { "categoryId": 2, "categoryName": "Category 2" },
                    { "categoryId": 3, "categoryName": "Category 3" },
                    { "categoryId": 4, "categoryName": "Category 4" },
                    { "categoryId": 5, "categoryName": "Category 5" },
                    { "categoryId": 6, "categoryName": "Category 6" },
                    { "categoryId": 7, "categoryName": "Category 7" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid blood group index provided.",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid blood group index"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "example": {
                  "error": "An unexpected error occurred."
                }
              }
            }
          }
        }
      }
    },
    "/api/public/products": {
      "get": {
        "summary": "Get all health products",
        "description": "Fetches all products from the Health collection",
        "operationId": "getAllHealthProducts",
        "tags": ["Nutrition"],
        "responses": {
          "200": {
            "description": "List of all health products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "5d51694902b2373622ff58a3"
                      },
                      "categories": { "type": "string", "example": "eggs" },
                      "weight": { "type": "number", "example": 100 },
                      "title": {
                        "type": "string",
                        "example": "Chicken egg (boiled)"
                      },
                      "calories": { "type": "number", "example": 160 },
                      "groupBloodNotAllowed": {
                        "type": "array",
                        "items": { "type": "boolean" },
                        "example": [null, true, false, false, false]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": { "error": "An unexpected error occurred." }
              }
            }
          }
        }
      }
    },
...
    "/api/private/{height}/{desiredWeight}/{age}/{bloodGroupIndex}/{currentWeight}": {
      "get": {
        "summary": "Get private recommended daily calories intake based on height, desiredWeight, age, currentWeight and returnes forbiden  categories of aliments to eat based on blood group",
        "description": "This endpoint calculates the recommended daily calorie intake based on height, desired weight, current weight, and age, and retrieves health categories based on the blood group index. It checks if the given blood group is allowed and fetches categories accordingly.",
        "operationId": "getPrivateCategoriesForBloodGroup",
        "parameters": [
          {
            "name": "height",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 170
            },
            "description": "The height of the individual in cm."
          },
          {
            "name": "desiredWeight",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 70
            },
            "description": "The desired weight of the individual in kg."
          },
          {
            "name": "age",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 30
            },
            "description": "The age of the individual."
          },
          {
            "name": "bloodGroupIndex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2
            },
            "description": "The blood group index for fetching the relevant categories."
          },
          {
            "name": "currentWeight",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 75
            },
            "description": "The current weight of the individual in kg."
          }
        ],
        "tags": ["Nutrition"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successfully retrieved daily calorie intake and categories for the blood group.",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "recommendedDailyCaloriesIntake": 2200,
                  "data": [
                    { "categoryId": 1, "categoryName": "Category 1" },
                    { "categoryId": 2, "categoryName": "Category 2" },
                    { "categoryId": 3, "categoryName": "Category 3" },
                    { "categoryId": 4, "categoryName": "Category 4" },
                    { "categoryId": 5, "categoryName": "Category 5" },
                    { "categoryId": 6, "categoryName": "Category 6" },
                    { "categoryId": 7, "categoryName": "Category 7" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid blood group index provided.",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid blood group index"
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "example": {
                  "error": "An unexpected error occurred."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}